Stats Script not updating immediately page loads
the main script has 4 default items in the array
the details are displayed on dom but not updated
in the stats section

i had to dig into the issue.
I used the debugger tool in chrome to go into the scripts
one after the other.

I noticed that script tag execute first and puts data into pubsub
But since stats script module is yet toload and no function placement
in the pubsub js, the data provided by the script js has no where to
run thereby making no changes to the stats section immediatley the page
loads.

i used debugger tool to look into each line of code to trouble shoot the
cause of the problem.
The dow arrow enters each line of code. the circle arrow enters function
and the up arrow exits function.
In all, the debugger section is cool tool to use.

Solution to the identified issue is to load the stats script first
because it is the one that plaves function into the pubsub for the data
coming from main script to use.
So i placed the stats script on top in html head before main script.